+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 
 ### Features
 
-* Change the priority of `request.meta.get('proxy')` and `pyppeteer_meta.get('proxy')`
 * Add `pretend` attribute for `PyppeteerRequest`, which can override `GERAPY_PYPPETEER_PRETEND`
 * Add support for `dict` format of `wait_for` attribute of `PyppeteerRequest`
+
+### Bug Fixes
+
+* Change the priority of `request.meta.get('proxy')` and `pyppeteer_meta.get('proxy')`
 
 ## 0.0.9 (2020-07-31)